requirements.txt
certifi==2019.3.9
chardet==3.0.4
Django==2.2.2
djangorestframework==3.9.4
docopt==0.6.2
idna==2.8
Pillow==6.0.0
pipreqs==0.4.9
psycopg2==2.8.3
pytz==2019.1
requests==2.22.0
sqlparse==0.3.0
urllib3==1.25.3
yarg==0.1.9
